Made it teleport to Erebus :)
Portal block under vanilla tools creative tab.
had to disable all the things...
Biomes exist from what the debug tells me - but no terrain gen so it is
just a void atm...

+++ src/main/java/erebus/world/genlayer/GenLayerErebus.java
@@ -36,6 +36,6 @@ public abstract class GenLayerErebus extends GenLayer {
 	public static byte getModdedBiomeSize(WorldType worldType, byte original) {
 		WorldTypeEvent.BiomeSize event = new WorldTypeEvent.BiomeSize(worldType, original);
 		MinecraftForge.TERRAIN_GEN_BUS.post(event);
-		return event.newSize;
+		return (byte) event.newSize;
 	}
 }
